infra/compliance-scanners/lambdas/process_document/requirements.txt
PyPDF2==3.0.1
langchain-pinecone==0.2.0
langchain-google-genai==2.0.8
langchain==0.3.0
pinecone-client==5.0.1